requirements.txt
fastapi
uvicorn
gunicorn
motor
redis>=5.0.0
python-dotenv
